requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
pdfplumber
pytesseract
pdf2image
scikit-learn
pillow
pydantic-settings
python-dotenv
python-multipart
SQLAlchemy>=2.0
psycopg2-binary
selenium
webdriver-manager
beautifulsoup4
requests
